Unspecified Task is handled as a Manual Task
Whenever a task is not specified in its type, it will get represented as manual task (a symbol with a hand). This behavior should be corrected. I already made the required changes locally, but I am no collaborator to this repository. So if you would like to add me as a collaborator I could push my changes and make a pull request, so that this issue gets fixed. I am also working on representation for message flows in this component. This file change is not sufficient, its okay to discard this request, but I would be glad if I could contribute to this repository.